requirements.txt
pinecone-client==2.2.2
sentence-transformers==2.2.2
langchain==0.0.228
# pydantic==2.0.2
python-dotenv==1.0.0
pydantic<2.0.0,>=1.8.2
# pydantic-settings==2.0.1
black==23.1.0
ruff==0.0.245
datasets==2.13.1
apache_beam==2.48.0
tiktoken==0.4.0
openai==0.27.8